Add flatpak autoconf patch to patches.txt
For switching flatpak builds over to using source + bsys builds,
the patch fixing the (odd?) default handling of the browser being in a
flatpak environment/sandbox (causing it to not load any of the LibreWolf
settings at all) would need to be added to the patches.txt, too.

It should not break anything else, as it is relatively specific to only
modify in-flatpak-behavior, as far as I understand it.